+++ lib/ui/attributes/ContentAttributesCategoryPanel.dart
@@ -17,21 +17,26 @@ class ContentAttributesCategoryPanel extends StatefulWidget {
   final ContentAttributes? contentAttributes;
   final List<ContentAttributeValue>? attributeValues;
   final LinkedHashSet<dynamic>? selection;
-  final bool multipleSelection;
   final bool filtersMode;
 
-  ContentAttributesCategoryPanel({this.attribute, this.contentAttributes, this.attributeValues, this.selection, this.multipleSelection = false, this.filtersMode = false});
+  ContentAttributesCategoryPanel({this.attribute, this.contentAttributes, this.attributeValues, this.selection, this.filtersMode = false});
 
   @override
   State<StatefulWidget> createState() => _ContentAttributesCategoryPanelState();
 
-  LinkedHashSet<String> get emptySelection => (attribute?.nullValue != null) ? LinkedHashSet<String>.from([attribute?.nullValue]) : LinkedHashSet<String>();
+  LinkedHashSet<dynamic> get emptySelection => (attribute?.nullValue != null) ? LinkedHashSet<dynamic>.from([attribute?.nullValue]) : LinkedHashSet<dynamic>();
 }
 
 class _ContentAttributesCategoryPanelState extends State<ContentAttributesCategoryPanel> {
 
   List<dynamic> _contentList = <dynamic>[];
   LinkedHashSet<dynamic> _selection = LinkedHashSet<dynamic>();
+
+  int get requirementsScope => widget.filtersMode ? contentAttributeRequirementsScopeFilter : contentAttributeRequirementsScopeCreate;
+
+  bool get singleSelection => widget.attribute?.isSingleSelection(requirementsScope) ?? true;
+  bool get multipleSelection => widget.attribute?.isMultipleSelection(requirementsScope) ?? false;
+  bool get multipleValueGroups => widget.attribute?.hasMultipleValueGroups ?? false;
 
   @override
   void initState() {
@@ -82,23 +87,25 @@ class _ContentAttributesCategoryPanelState extends State<ContentAttributesCatego
     super.dispose();
   }
 
+  // Widget
+
   @override
   Widget build(BuildContext context) {
     String? title = widget.attribute?.displayTitle;
 
     List<Widget> actions = <Widget>[];
 
-    if ((0 < (widget.attributeValues?.length ?? 0)) && !widget.filtersMode && (widget.attribute?.isSingleSelection ?? false) && !DeepCollectionEquality().equals(_selection, widget.emptySelection)) {
-      actions.add(_buildHeaderBarButton(
-        title:  Localization().getStringEx('dialog.clear.title', 'Clear'),
-        onTap: _onTapClear,
-      ));
-    }
 
-    if (widget.multipleSelection && !DeepCollectionEquality().equals(widget.selection ?? LinkedHashSet<dynamic>(), _selection)) {
+    if ((multipleSelection || multipleValueGroups) && !DeepCollectionEquality().equals(widget.selection ?? LinkedHashSet<dynamic>(), _selection)) {
       actions.add(_buildHeaderBarButton(
         title:  Localization().getStringEx('dialog.apply.title', 'Apply'),
         onTap: _onTapApply,
+      ));
+    }
+    else if ((0 < (widget.attributeValues?.length ?? 0)) && !DeepCollectionEquality().equals(_selection, widget.emptySelection)) {
+      actions.add(_buildHeaderBarButton(
+        title:  Localization().getStringEx('dialog.clear.title', 'Clear'),
+        onTap: _onTapClear,
       ));
     }
 
@@ -186,13 +193,18 @@ class _ContentAttributesCategoryPanelState extends State<ContentAttributesCatego
   Widget _buildAttributeValueWidget(ContentAttributeValue attributeValue) {
     bool isSelected = _selection.contains(attributeValue.value);
     String? imageAsset = (attributeValue.value != null) ?
-      (widget.multipleSelection ?
+      (multipleSelection ?
         (isSelected ? "check-box-filled" : "box-outline-gray") :
         (isSelected ? "check-circle-filled" : "circle-outline-gray")
       ) : null;
+    
     String? title = StringUtils.isNotEmpty(attributeValue.label) ?
       widget.attribute?.displayString(attributeValue.label) :
       Localization().getStringEx('panel.content.attributes.button.clear.title', 'Clear');
+    
+    String? info = StringUtils.isNotEmpty(attributeValue.info) ?
+      widget.attribute?.displayString(attributeValue.info) : null;
+
     TextStyle? textStyle = (attributeValue.value != null) ?
       Styles().textStyles?.getTextStyle(isSelected ? "widget.group.dropdown_button.item.selected" : "widget.group.dropdown_button.item.not_selected") :
       Styles().textStyles?.getTextStyle("widget.label.regular.thin");
@@ -206,8 +218,16 @@ class _ContentAttributesCategoryPanelState extends State<ContentAttributesCatego
               Text(title ?? attributeValue.label ?? '', overflow: TextOverflow.ellipsis, style: textStyle,),
             )
           ),
-          
-          Styles().images?.getImage(imageAsset, excludeFromSemantics: true) ?? Container()
+
+          Row(mainAxisSize: MainAxisSize.min, children: <Widget>[
+            Visibility(visible: StringUtils.isNotEmpty(info), child:
+              Padding(padding: const EdgeInsets.only(right: 8), child:
+                Text(info ?? attributeValue.info ?? '', overflow: TextOverflow.ellipsis, style: textStyle,),
+              )
+            ),
+            
+            Styles().images?.getImage(imageAsset, excludeFromSemantics: true) ?? Container()
+          ]),
         ]),
       )
     );
@@ -229,11 +249,11 @@ class _ContentAttributesCategoryPanelState extends State<ContentAttributesCatego
       _selection.clear();
     }
 
-    if (widget.attribute?.requirements?.hasScope(widget.filtersMode ? contentAttributeRequirementsScopeFilter : contentAttributeRequirementsScopeCreate) ?? false) {
+    if (widget.attribute?.requirements?.hasScope(requirementsScope) ?? false) {
       widget.attribute?.validateAttributeValuesSelection(_selection);
     }
 
-    if (widget.multipleSelection || widget.filtersMode) {
+    if (multipleSelection || multipleValueGroups) {
       setStateIfMounted(() { });
     }
     else {
@@ -247,8 +267,8 @@ class _ContentAttributesCategoryPanelState extends State<ContentAttributesCatego
   }
 
   void _onTapClear() {
-    _selection = widget.emptySelection;
-    if (widget.multipleSelection) {
+    _selection = LinkedHashSet<dynamic>.from(widget.emptySelection);
+    if (multipleSelection || multipleValueGroups) {
       setStateIfMounted(() {});
     }
     else {