+++ CHANGELOG.md
@@ -6,10 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [5.0.10] - 2023-06-16
+### Added
+- Added Event Time filter and related attribute updates [#3300](https://github.com/rokwire/illinois-app/issues/3300).
+
 ## [5.0.9] - 2023-06-15
+### Added
+- Added Event Types filter and related attribute updates [#3300](https://github.com/rokwire/illinois-app/issues/3300).
 ### Changed
 - Export button text styles to styles.json [#2932](https://github.com/rokwire/illinois-app/issues/2932).
-- Added Event Types filter and related attribute updates [#3300](https://github.com/rokwire/illinois-app/issues/3300).
 
 ## [5.0.8] - 2023-06-12
 ### Added